pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>=42", # At least v42 of setuptools required.
  "versioningit",
]

[tool.versioningit.vcs]
method = "git"
default-tag = "0.0.0"

[project]
name = "influxio"
description = "Import and export data into/from InfluxDB"
readme = "README.rst"
keywords = [
  "export",
  "ilp",
  "import",
  "influx",
  "influxdb",
  "line-protocol",
  "lineprotocol",
  "rdbms",
  "sql",
  "sqlalchemy",
  "timeseries",
  "timeseries-data",
  "transfer",
]
license = { text = "MIT" }
authors = [
  { name = "Andreas Motl", email = "andreas.motl@panodata.org" },
  { name = "Richard Pobering", email = "richard.pobering@panodata.org" },
]
requires-python = ">=3.8"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Customer Service",
  "Intended Audience :: Developers",
  "Intended Audience :: Education",
  "Intended Audience :: End Users/Desktop",
  "Intended Audience :: Information Technology",
  "Intended Audience :: Manufacturing",
  "Intended Audience :: Science/Research",
  "Intended Audience :: System Administrators",
  "Intended Audience :: Telecommunications Industry",
  "License :: OSI Approved :: MIT License",
  "Operating System :: MacOS :: MacOS X",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "Operating System :: Unix",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Communications",
  "Topic :: Database",
  "Topic :: Documentation",
  "Topic :: Education",
  "Topic :: Office/Business",
  "Topic :: Scientific/Engineering",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development :: Testing",
  "Topic :: Software Development :: Version Control",
  "Topic :: System :: Archiving",
  "Topic :: System :: Systems Administration",
  "Topic :: Text Processing",
  "Topic :: Utilities",
]
dynamic = [
  "version",
]
dependencies = [
  "click<9",
  "colorama<1",
  "crate[sqlalchemy]",
  "cratedb-toolkit",
  "dask",
  'importlib-metadata; python_version <= "3.9"',
  "influxdb-client[ciso]<2",
  "line-protocol-parser<2",
  "pandas<2.2",
  "psycopg2-binary<3",
  "SQLAlchemy-Utils<0.42",
  "yarl<2",
]
[project.optional-dependencies]
develop = [
  "black<24",
  "mypy==1.6.1",
  "poethepoet<0.25",
  "pyproject-fmt<1.5",
  "ruff==0.1.3",
  "validate-pyproject<0.16",
]
release = [
  "build<2",
  "twine<5",
]
test = [
  "pytest<8",
  "pytest-cov<5",
]
[project.urls]
changelog = "https://github.com/daq-tools/influxio/blob/main/CHANGES.rst"
documentation = "https://github.com/daq-tools/influxio"
homepage = "https://github.com/daq-tools/influxio"
repository = "https://github.com/daq-tools/influxio"
[project.scripts]
influxio = "influxio.cli:cli"

[tool.black]
line-length = 120

[tool.coverage.run]
branch = false
omit = [
  "tests/*",
]
source = ["influxio"]

[tool.coverage.report]
fail_under = 0
show_missing = true

[tool.mypy]
packages = ["influxio"]
install_types = true
ignore_missing_imports = true
implicit_optional = true
non_interactive = true

[tool.pytest.ini_options]
addopts = "-rA --verbosity=3 --cov --cov-report=term-missing --cov-report=xml"
minversion = "2.0"
log_level = "DEBUG"
log_cli_level = "DEBUG"
log_format = "%(asctime)-15s [%(name)-24s] %(levelname)-8s: %(message)s"
testpaths = [
  "influxio",
  "tests",
]
xfail_strict = true
markers = [
  "examples",
  "slow",
]

[tool.ruff]
line-length = 120

select = [
  # Bandit
  "S",
  # Bugbear
  "B",
  # Builtins
  "A",
  # comprehensions
  "C4",
  # eradicate
  "ERA",
  # flake8-2020
  "YTT",
  # isort
  "I",
  # pandas-vet
  "PD",
  # print
  "T20",
  # Pycodestyle
  "E",
  "W",
  # Pyflakes
  "F",
  # return
  "RET",
]

extend-ignore = [
  # zip() without an explicit strict= parameter
  "B905",
  # df is a bad variable name. Be kinder to your future self.
  "PD901",
  # Unnecessary variable assignment before `return` statement
  "RET504",
  # Unnecessary `elif` after `return` statement
  "RET505",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]  # Use of `assert` detected
"influxio/util/report.py" = ["T201"]

[tool.setuptools.packages.find]
namespaces = false


# ===================
# Tasks configuration
# ===================

[tool.poe.tasks]

check = [
  "lint",
  "test",
]

format = [
  { cmd = "black ." },
  # Configure Ruff not to auto-fix (remove!) unused variables (F841) and `print` statements (T201).
  { cmd = "ruff --fix --ignore=ERA --ignore=F401 --ignore=F841 --ignore=T20 ." },
  { cmd = "pyproject-fmt pyproject.toml" },
]

lint = [
  { cmd = "ruff ." },
  { cmd = "black --check ." },
  { cmd = "validate-pyproject pyproject.toml" },
  # { cmd = "mypy" },
]

release = [
  { cmd = "python -m build" },
  { cmd = "twine upload dist/*" },
]

test = { cmd = "pytest" }
